+++ otterdog/models/branch_protection_rule.py
@@ -10,7 +10,7 @@ from __future__ import annotations
 
 import dataclasses
 import re
-from typing import Any, Optional, Iterator
+from typing import Any, Optional
 
 from jsonbender import bend, S, OptionalS, Forall, K  # type: ignore
 
@@ -66,15 +66,13 @@ class BranchProtectionRule(ModelObject):
         return "branch_protection_rule"
 
     def validate(self, context: ValidationContext, parent_object: Any) -> None:
-        repo_name: str = parent_object.name
-
         # when requires_approving_reviews is false, issue a warning if dependent settings
         # are still set to non default values.
 
         if self.requires_approving_reviews is False:
             if is_set_and_valid(self.required_approving_review_count):
                 context.add_failure(FailureType.WARNING,
-                                    f"branch_protection_rule[repo=\"{repo_name}\", pattern=\"{self.pattern}\"] has"
+                                    f"{self.get_model_header(parent_object)} has"
                                     f" 'requires_approving_reviews' disabled but 'required_approving_review_count' "
                                     f"is set to '{self.required_approving_review_count}', setting will be ignored.")
 
@@ -84,7 +82,7 @@ class BranchProtectionRule(ModelObject):
                         "restricts_review_dismissals"]:
                 if self.__getattribute__(key) is True:
                     context.add_failure(FailureType.WARNING,
-                                        f"branch_protection_rule[repo=\"{repo_name}\", pattern=\"{self.pattern}\"] has"
+                                        f"{self.get_model_header(parent_object)} has"
                                         f" 'requires_approving_reviews' disabled but '{key}' "
                                         f"is enabled, setting will be ignored.")
 
@@ -92,7 +90,7 @@ class BranchProtectionRule(ModelObject):
                 value = self.__getattribute__(key)
                 if not is_unset(value) and len(value) > 0:
                     context.add_failure(FailureType.WARNING,
-                                        f"branch_protection_rule[repo=\"{repo_name}\", pattern=\"{self.pattern}\"] has"
+                                        f"{self.get_model_header(parent_object)} has"
                                         f" 'requires_approving_reviews' disabled but '{key}' "
                                         f"is set to {value}, setting will be ignored.")
 
@@ -101,7 +99,7 @@ class BranchProtectionRule(ModelObject):
         if self.requires_approving_reviews is True and not is_unset(required_approving_review_count):
             if required_approving_review_count is None or required_approving_review_count < 0:
                 context.add_failure(FailureType.ERROR,
-                                    f"branch_protection_rule[repo=\"{repo_name}\", pattern=\"{self.pattern}\"] has"
+                                    f"{self.get_model_header(parent_object)} has"
                                     f" 'requires_approving_reviews' enabled but 'required_approving_review_count' "
                                     f"is not set (must be null or a non negative number).")
 
@@ -111,7 +109,7 @@ class BranchProtectionRule(ModelObject):
                 is_set_and_valid(review_dismissal_allowances) and \
                 len(review_dismissal_allowances) > 0:
             context.add_failure(FailureType.WARNING,
-                                f"branch_protection_rule[repo=\"{repo_name}\", pattern=\"{self.pattern}\"] has"
+                                f"{self.get_model_header(parent_object)} has"
                                 f" 'restricts_review_dismissals' disabled but "
                                 f"'review_dismissal_allowances' is set to {self.review_dismissal_allowances}, "
                                 f"setting will be ignored.")
@@ -122,7 +120,7 @@ class BranchProtectionRule(ModelObject):
                 is_set_and_valid(bypass_force_push_allowances) and \
                 len(bypass_force_push_allowances) > 0:
             context.add_failure(FailureType.WARNING,
-                                f"branch_protection_rule[repo=\"{repo_name}\", pattern=\"{self.pattern}\"] has"
+                                f"{self.get_model_header(parent_object)} has"
                                 f" 'allows_force_pushes' enabled but "
                                 f"'bypass_force_push_allowances' is set to {self.bypass_force_push_allowances}, "
                                 f"setting will be ignored.")
@@ -133,7 +131,7 @@ class BranchProtectionRule(ModelObject):
                 is_set_and_valid(required_status_checks) and \
                 len(required_status_checks) > 0:
             context.add_failure(FailureType.WARNING,
-                                f"branch_protection_rule[repo=\"{repo_name}\", pattern=\"{self.pattern}\"] has"
+                                f"{self.get_model_header(parent_object)} has"
                                 f" 'requires_status_checks' disabled but "
                                 f"'required_status_checks' is set to {self.required_status_checks}, "
                                 f"setting will be ignored.")
@@ -143,7 +141,7 @@ class BranchProtectionRule(ModelObject):
                 is_set_and_valid(self.required_deployment_environments) and \
                 len(self.required_deployment_environments) > 0:
             context.add_failure(FailureType.WARNING,
-                                f"branch_protection_rule[repo=\"{repo_name}\", pattern=\"{self.pattern}\"] has"
+                                f"{self.get_model_header(parent_object)} has"
                                 f" 'requires_deployments' disabled but "
                                 f"'required_deployment_environments' is set to {self.required_deployment_environments}"
                                 f", setting will be ignored.")
@@ -181,16 +179,13 @@ class BranchProtectionRule(ModelObject):
     def include_field_for_patch_computation(self, field: dataclasses.Field) -> bool:
         return True
 
-    def get_model_objects(self) -> Iterator[tuple[ModelObject, ModelObject]]:
-        yield from []
-
     @classmethod
     def from_model_data(cls, data: dict[str, Any]) -> BranchProtectionRule:
         mapping = {k: OptionalS(k, default=UNSET) for k in map(lambda x: x.name, cls.all_fields())}
         return cls(**bend(mapping, data))
 
     @classmethod
-    def from_provider_data(cls, data: dict[str, Any]) -> BranchProtectionRule:
+    def from_provider_data(cls, org_id: str, data: dict[str, Any]) -> BranchProtectionRule:
         mapping = {k: OptionalS(snake_to_camel_case(k), default=UNSET) for k in map(lambda x: x.name, cls.all_fields())}
 
         def transform_app(x):