+++ otterdog/models/webhook.py
@@ -10,11 +10,11 @@ from __future__ import annotations
 
 import abc
 import dataclasses
-from typing import Any, Optional, Iterator
+from typing import Any, Optional
 
 from jsonbender import bend, S, OptionalS  # type: ignore
 
-from otterdog.models import ModelObject, ValidationContext, FailureType
+from otterdog.models import ModelObject
 from otterdog.providers.github import Github
 from otterdog.utils import UNSET, is_unset, is_set_and_valid
 
@@ -40,13 +40,6 @@ class Webhook(ModelObject, abc.ABC):
     def get_all_urls(self) -> list[str]:
         return [self.url] + self.aliases
 
-    def validate(self, context: ValidationContext, parent_object: Any) -> None:
-        if self.has_dummy_secret():
-            context.add_failure(FailureType.WARNING,
-                                f"organization webhook with url '{self.url}' will be skipped during processing:\n"
-                                f"webhook has a secret set, but only a dummy secret '{self.secret}' is provided in "
-                                f"the configuration.")
-
     def has_dummy_secret(self) -> bool:
         if is_set_and_valid(self.secret) and all(ch == '*' for ch in self.secret):  # type: ignore
             return True
@@ -61,16 +54,13 @@ class Webhook(ModelObject, abc.ABC):
     def include_field_for_patch_computation(self, field: dataclasses.Field) -> bool:
         return True
 
-    def get_model_objects(self) -> Iterator[tuple[ModelObject, ModelObject]]:
-        yield from []
-
     @classmethod
     def from_model_data(cls, data: dict[str, Any]):
         mapping = {k: OptionalS(k, default=UNSET) for k in map(lambda x: x.name, cls.all_fields())}
         return cls(**bend(mapping, data))
 
     @classmethod
-    def from_provider_data(cls, data: dict[str, Any]):
+    def from_provider_data(cls, org_id: str, data: dict[str, Any]):
         mapping = {k: OptionalS(k, default=UNSET) for k in map(lambda x: x.name, cls.all_fields())}
         mapping.update(
             {