+++ otterdog/operations/show_live_operation.py
@@ -6,23 +6,23 @@
 # SPDX-License-Identifier: MIT
 # *******************************************************************************
 
-import os
-
 from colorama import Style, Fore
 
 from otterdog.config import OrganizationConfig
 from otterdog.models.github_organization import GitHubOrganization
-from otterdog.utils import print_error
+from otterdog.providers.github import Github
+from otterdog.utils import print_error, print_warn
 
 from . import Operation
 
 
-class ShowOperation(Operation):
-    def __init__(self):
+class ShowLiveOperation(Operation):
+    def __init__(self, no_web_ui: bool):
         super().__init__()
+        self.no_web_ui = no_web_ui
 
     def pre_execute(self) -> None:
-        self.printer.println(f"Showing resources defined in configuration '{self.config.config_file}'")
+        self.printer.println("Showing live resources for organizations")
 
     def execute(self, org_config: OrganizationConfig) -> int:
         github_id = org_config.github_id
@@ -33,21 +33,24 @@ class ShowOperation(Operation):
         self.printer.level_up()
 
         try:
-            org_file_name = jsonnet_config.org_config_file
-
-            if not os.path.exists(org_file_name):
-                print_error(f"configuration file '{org_file_name}' does not yet exist, run fetch first")
-                return 1
-
             try:
-                organization = \
-                    GitHubOrganization.load_from_file(github_id,
-                                                      org_file_name,
-                                                      self.config,
-                                                      False)
-            except RuntimeError as ex:
-                print_error(f"failed to load configuration: {str(ex)}")
+                credentials = self.config.get_credentials(org_config)
+            except RuntimeError as e:
+                print_error(f"invalid credentials\n{str(e)}")
                 return 1
+
+            gh_client = Github(credentials)
+
+            if self.no_web_ui is True:
+                print_warn("the Web UI will not be queried as '--no-web-ui' has been specified, "
+                           "the resulting config will be incomplete")
+
+            organization = \
+                GitHubOrganization.load_from_provider(github_id,
+                                                      jsonnet_config,
+                                                      gh_client,
+                                                      self.no_web_ui,
+                                                      self.printer)
 
             for model_object, parent_object in organization.get_model_objects():
                 self.printer.println()