+++ otterdog/providers/github/graphql.py
@@ -112,12 +112,12 @@ class GraphQLClient:
         else:
             raise RuntimeError(f"failed to update branch protection rule '{rule_pattern}'")
 
-    def add_branch_protection_rule(self, org_id: str, repo_name: str, repo_id: str, data: dict[str, Any]) -> None:
+    def add_branch_protection_rule(self, org_id: str, repo_name: str, repo_node_id: str, data: dict[str, Any]) -> None:
         rule_pattern = data["pattern"]
         utils.print_debug(f"creating branch_protection_rule with pattern '{rule_pattern}' "
                           f"for repo '{org_id}/{repo_name}'")
 
-        data["repositoryId"] = repo_id
+        data["repositoryId"] = repo_node_id
         variables = {"ruleInput": data}
 
         query = """mutation($ruleInput: CreateBranchProtectionRuleInput!) {